requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
openai==1.30.1
chromadb==0.4.22
tavily-python==0.3.5
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn[standard]==0.27.1
aiohttp==3.9.3
requests==2.31.0
regex==2023.10.3
mcp[cli]==1.2.0
sse-starlette==1.6.5
dspy-ai==2.3.1  